update(authentication): fixed the JWT authentication for verifying the owner of resources
integrate(Google-oauth): added google oauth for login

update(database): added user from other providers

+++ backend/node/controllers/comment.controller.js
@@ -23,6 +23,13 @@ class CommentController {
     }
 
     static async deleteComment(req, res) {
+        const userId = req.user.id; // Assuming AuthMiddleware sets req.user
+        const { authorId } = req.body; // Assuming authorId is sent in the body
+
+        if (authorId && userId !== authorId) {
+            return res.status(403).json({ error: "You are not authorized to delete this comment" });
+        }
+
         try {
             await Comment.deleteComment(req.params.comment_id);
             res.status(200).json({ message: "Comment deleted successfully"});
